requirements.txt
python-telegram-bot==20.6
firebase-admin==6.4.0

